+++ src/SME.SGP.Dados/Repositorios/RepositorioWfAprovacaoNotaFechamento.cs
@@ -76,7 +76,7 @@ namespace SME.SGP.Dados.Repositorios
         {
             var query = @"select ft.turma_id as TurmaId, t.ano_letivo as AnoLetivo, fa.fechamento_turma_disciplina_id as FechamentoTurmaDisciplinaId, pe.bimestre as Bimestre, 
                                 fa.aluno_codigo as CodigoAluno, fn.nota as NotaAnterior, coalesce(cc.descricao_infantil, cc.descricao_sgp, cc.descricao) as ComponenteCurricularDescricao, 
-                                cc.eh_regencia as ComponenteCurricularEhRegencia, cc.eh_regencia as ComponenteCurricularEhRegencia,
+                                cc.eh_regencia as ComponenteCurricularEhRegencia, 
                                 cc.permite_lancamento_nota LancaNota, fn.conceito_id as ConceitoAnteriorId, wanf.* 
                             from wf_aprovacao_nota_fechamento wanf 
                              join fechamento_nota fn on fn.id = wanf.fechamento_nota_id
