+++ src/SME.SGP.Infra/Dtos/FechamentoNota/WfAprovacaoNotaFechamentoTurmaDto.cs
@@ -8,7 +8,6 @@ namespace SME.SGP.Infra
         public long TurmaId { get; set; }
         public long FechamentoTurmaDisciplinaId { get; set; }
         public string ComponenteCurricularDescricao { get; set; }
-        public bool ComponenteCurricularEhRegencia { get; set; }
         public FechamentoNota FechamentoNota { get; set; }
         public string CodigoAluno { get; set; }
         public double? NotaAnterior { get; set; }
